Extract route helpers to separate file
